Move RoarDataTable test specs into component dir
