+++ src/components/RoarDataTable/index.js
@@ -1,0 +1,1 @@
+export { default } from './RoarDataTable.vue';
