+++ src/pages/ProgressReport.vue
@@ -187,7 +187,7 @@ import { getTitle } from '@/helpers/query/administrations';
 import { setBarChartData, setBarChartOptions } from '@/helpers/plotting';
 import { APP_ROUTES } from '@/constants/routes';
 import { SINGULAR_ORG_TYPES } from '@/constants/orgTypes';
-import RoarDataTable from '@/components/RoarDataTable.vue';
+import RoarDataTable from '@/components/RoarDataTable';
 
 const router = useRouter();
 const authStore = useAuthStore();
