+++ src/pages/ScoreReport.vue
@@ -310,7 +310,7 @@ import {
   includedValidityFlags,
 } from '@/helpers/reports';
 import FilterBar from '@/components/slots/FilterBar.vue';
-import RoarDataTable from '@/components/RoarDataTable.vue';
+import RoarDataTable from '@/components/RoarDataTable';
 import { APP_ROUTES } from '@/constants/routes';
 import { SINGULAR_ORG_TYPES } from '@/constants/orgTypes';
